fix: correct three genuine return-type bugs surfaced by phpstan
- StatementRepository::getTotalStatementsForDateRange() returned a query
  Builder under a `: float` declaration (a guaranteed TypeError on any
  call). It has no callers; complete the query with ->count() and type it
  `: int` ("total statements" = how many in the range). Add a regression
  test.
- ReferrerTestExport::getPrice(): string returned $referrerTest->price
  (float|null) in the PANEL branch — a null would TypeError. Cast to string.
- TATService::calcOnTimePct(): ?int returned round() (float); wrap in
  (int) so the value matches the declared type.

Drop the three resolved entries from the phpstan baseline. The remaining
16 return.type baseline entries are confirmed false positives (Eloquent
returns the generic Model/Contract while the runtime value is the concrete
subclass, plus a Carbon\Carbon-vs-Illuminate\Support\Carbon covariance
nitpick) — none are runtime bugs.

+++ app/Domains/Reception/Services/TATService.php
@@ -328,6 +328,6 @@ class TATService
             return $publishedAt && $publishedAt->lte($deadline);
         })->count();
 
-        return round(($onTime / $total) * 100);
+        return (int) round(($onTime / $total) * 100);
     }
 }
